requirements.txt
django-model-utils
tox
